Update CR and minluck formula to account for floating-point errors
Dave has confirmed that the minluck of Zealous Academic is in fact 86,
not 85. My theory for this is that it's because of floating point errors
in MouseHunt (and indeed Dave hinted that the explanation was a silly
quirk, not an originally intended part of the formula). This theory is
described in further detail at https://mh.beeejk.net/minluck (and has
been discussed in Discord).

The fixes for CRE are pretty simple; we just have to remove the care the
code was taking to avoid floating-point errors! And then for minluck we
need to check if the error is happening, and if so increment by 1. I
also added to the README an abbreviated explanation and a link to the
full explanation.

Let me know if there's anywhere other than the README that I should
surface the change to users.